renamed comparison to logic to keep consistency with routines
