+++ chainerx_cc/chainerx/native/CMakeLists.txt
@@ -18,7 +18,7 @@ add_library(chainerx_native STATIC
     native_device/binary.cc
     native_device/conv.cc
     native_device/copy.cc
-    native_device/comparison.cc
+    native_device/logic.cc
     native_device/dot.cc
     native_device/exp_log.cc
     native_device/fill.cc
