+++ chainerx_cc/chainerx/native/native_device/misc.cc
@@ -6,7 +6,6 @@
 #include "chainerx/array.h"
 #include "chainerx/device.h"
 #include "chainerx/dtype.h"
-#include "chainerx/kernels/logic.h"
 #include "chainerx/kernels/math.h"
 #include "chainerx/kernels/misc.h"
 #include "chainerx/native/elementwise.h"
@@ -29,54 +28,6 @@ CHAINERX_NATIVE_REGISTER_ELTWISE_DTYPE_UNARY_KERNEL(SignKernel, { out = chainerx
 CHAINERX_NATIVE_REGISTER_ELTWISE_FLOAT_UNARY_KERNEL(CeilKernel, { out = chainerx::Ceil(x); });
 
 CHAINERX_NATIVE_REGISTER_ELTWISE_FLOAT_UNARY_KERNEL(FloorKernel, { out = chainerx::Floor(x); });
-
-class NativeIsNanKernel : public IsNanKernel {
-public:
-    void Call(const Array& x, const Array& out) override {
-        x.device().CheckDevicesCompatible(x, out);
-        VisitDtype(x.dtype(), [&](auto pt) {
-            using T = typename decltype(pt)::type;
-            struct Impl {
-                void operator()(int64_t /*i*/, T x, bool& out) { out = chainerx::IsNan(x); }
-            };
-            Elementwise<const T, bool>(Impl{}, x, out);
-        });
-    }
-};
-
-CHAINERX_NATIVE_REGISTER_KERNEL(IsNanKernel, NativeIsNanKernel);
-
-class NativeIsInfKernel : public IsInfKernel {
-public:
-    void Call(const Array& x, const Array& out) override {
-        x.device().CheckDevicesCompatible(x, out);
-        VisitDtype(x.dtype(), [&](auto pt) {
-            using T = typename decltype(pt)::type;
-            struct Impl {
-                void operator()(int64_t /*i*/, T x, bool& out) { out = chainerx::IsInf(x); }
-            };
-            Elementwise<const T, bool>(Impl{}, x, out);
-        });
-    }
-};
-
-CHAINERX_NATIVE_REGISTER_KERNEL(IsInfKernel, NativeIsInfKernel);
-
-class NativeIsFiniteKernel : public IsFiniteKernel {
-public:
-    void Call(const Array& x, const Array& out) override {
-        x.device().CheckDevicesCompatible(x, out);
-        VisitDtype(x.dtype(), [&](auto pt) {
-            using T = typename decltype(pt)::type;
-            struct Impl {
-                void operator()(int64_t /*i*/, T x, bool& out) { out = !(chainerx::IsInf(x) || chainerx::IsNan(x)); }
-            };
-            Elementwise<const T, bool>(Impl{}, x, out);
-        });
-    }
-};
-
-CHAINERX_NATIVE_REGISTER_KERNEL(IsFiniteKernel, NativeIsFiniteKernel);
 
 class NativeIfLessElseASSAKernel : public IfLessElseASSAKernel {
 public:
